message test [nfc]: Move ReactionEvent tests from msglist test file
