+++ test/model/message_test.dart
@@ -7,7 +7,9 @@ import 'package:zulip/model/narrow.dart';
 import 'package:zulip/model/store.dart';
 
 import '../api/fake_api.dart';
+import '../api/model/model_checks.dart';
 import '../example_data.dart' as eg;
+import '../stdlib_checks.dart';
 import 'message_list_test.dart';
 import 'test_store.dart';
 
@@ -49,7 +51,6 @@ void main() {
   /// Perform the initial message fetch for [messageList].
   ///
   /// The test case must have already called [prepare] to initialize the state.
-  // ignore: unused_element
   Future<void> prepareMessages(
     List<Message> messages, {
     bool foundOldest = false,
@@ -156,6 +157,89 @@ void main() {
       final newMessage = eg.streamMessage(id: 1, content: '<p>bar</p>');
       await store.handleEvent(MessageEvent(id: 1, message: newMessage));
       check(store.messages).deepEquals({1: newMessage});
+    });
+  });
+
+  group('handleReactionEvent', () {
+    ReactionEvent mkEvent(Reaction reaction, ReactionOp op, int messageId) {
+      return ReactionEvent(
+        id: 1,
+        op: op,
+        emojiName: reaction.emojiName,
+        emojiCode: reaction.emojiCode,
+        reactionType: reaction.reactionType,
+        userId: reaction.userId,
+        messageId: messageId,
+      );
+    }
+
+    test('add reaction', () async {
+      final originalMessage = eg.streamMessage(reactions: []);
+      await prepare();
+      await prepareMessages([originalMessage]);
+      final message = store.messages.values.single;
+
+      await store.handleEvent(
+        mkEvent(eg.unicodeEmojiReaction, ReactionOp.add, originalMessage.id));
+      checkNotifiedOnce();
+      check(store.messages).values.single
+        ..identicalTo(message)
+        ..reactions.isNotNull().jsonEquals([eg.unicodeEmojiReaction]);
+    });
+
+    test('add reaction; message is unknown', () async {
+      final someMessage = eg.streamMessage(reactions: []);
+      await prepare();
+      await prepareMessages([someMessage]);
+      await store.handleEvent(
+        mkEvent(eg.unicodeEmojiReaction, ReactionOp.add, 1000));
+      checkNotNotified();
+      check(store.messages).values.single
+        .reactions.isNull();
+    });
+
+    test('remove reaction', () async {
+      final eventReaction = Reaction(reactionType: ReactionType.unicodeEmoji,
+        emojiName: 'wave',          emojiCode: '1f44b', userId: 1);
+
+      // Same emoji, different user. Not to be removed.
+      final reaction2 = Reaction(reactionType: ReactionType.unicodeEmoji,
+        emojiName: 'wave',          emojiCode: '1f44b', userId: 2);
+
+      // Same user, different emoji. Not to be removed.
+      final reaction3 = Reaction(reactionType: ReactionType.unicodeEmoji,
+        emojiName: 'working_on_it', emojiCode: '1f6e0', userId: 1);
+
+      // Same user, same emojiCode, different emojiName. To be removed: servers
+      // key on user, message, reaction type, and emoji code, but not emoji name.
+      // So we mimic that behavior; see discussion:
+      //   https://github.com/zulip/zulip-flutter/pull/256#discussion_r1284865099
+      final reaction4 = Reaction(reactionType: ReactionType.unicodeEmoji,
+        emojiName: 'hello',         emojiCode: '1f44b', userId: 1);
+
+      final originalMessage = eg.streamMessage(
+        reactions: [reaction2, reaction3, reaction4]);
+      await prepare();
+      await prepareMessages([originalMessage]);
+      final message = store.messages.values.single;
+
+      await store.handleEvent(
+        mkEvent(eventReaction, ReactionOp.remove, originalMessage.id));
+      checkNotifiedOnce();
+      check(store.messages).values.single
+        ..identicalTo(message)
+        ..reactions.isNotNull().jsonEquals([reaction2, reaction3]);
+    });
+
+    test('remove reaction; message is unknown', () async {
+      final someMessage = eg.streamMessage(reactions: [eg.unicodeEmojiReaction]);
+      await prepare();
+      await prepareMessages([someMessage]);
+      await store.handleEvent(
+        mkEvent(eg.unicodeEmojiReaction, ReactionOp.remove, 1000));
+      checkNotNotified();
+      check(store.messages).values.single
+        .reactions.isNotNull().jsonEquals([eg.unicodeEmojiReaction]);
     });
   });
 }
